backend/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
pyotp==2.9.0
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
